+++ sample/src/com/viewpagerindicator/sample/SampleTabsStyled.java
@@ -2,11 +2,13 @@ package com.viewpagerindicator.sample;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentActivity;
 import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentPagerAdapter;
 import android.support.v4.view.ViewPager;
 import com.viewpagerindicator.TabPageIndicator;
 
-public class SampleTabsStyled extends BaseSampleActivity {
+public class SampleTabsStyled extends FragmentActivity {
     private static final String[] CONTENT = new String[] { "Recent", "Artists", "Albums", "Songs", "Playlists", "Genres" };
 
     @Override
@@ -14,33 +16,33 @@ public class SampleTabsStyled extends BaseSampleActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.simple_tabs);
 
-        mAdapter = new GoogleMusicAdapter(getSupportFragmentManager());
+        FragmentPagerAdapter adapter = new GoogleMusicAdapter(getSupportFragmentManager());
 
-        mPager = (ViewPager)findViewById(R.id.pager);
-        mPager.setAdapter(mAdapter);
+        ViewPager pager = (ViewPager)findViewById(R.id.pager);
+        pager.setAdapter(adapter);
 
-        mIndicator = (TabPageIndicator)findViewById(R.id.indicator);
-        mIndicator.setViewPager(mPager);
+        TabPageIndicator indicator = (TabPageIndicator)findViewById(R.id.indicator);
+        indicator.setViewPager(pager);
     }
 
-    class GoogleMusicAdapter extends TestFragmentAdapter {
+    class GoogleMusicAdapter extends FragmentPagerAdapter {
         public GoogleMusicAdapter(FragmentManager fm) {
             super(fm);
         }
 
         @Override
         public Fragment getItem(int position) {
-            return TestFragment.newInstance(SampleTabsStyled.CONTENT[position % SampleTabsStyled.CONTENT.length]);
-        }
-
-        @Override
-        public int getCount() {
-            return SampleTabsStyled.CONTENT.length;
+            return TestFragment.newInstance(CONTENT[position % CONTENT.length]);
         }
 
         @Override
         public CharSequence getPageTitle(int position) {
-            return SampleTabsStyled.CONTENT[position % SampleTabsStyled.CONTENT.length].toUpperCase();
+            return CONTENT[position % CONTENT.length].toUpperCase();
+        }
+
+        @Override
+        public int getCount() {
+            return CONTENT.length;
         }
     }
 }